refactor: Extract save/load state related function into module
